% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.updater.R
\name{beta.updater}
\alias{beta.updater}
\title{Updating Parameter Values within IRLS}
\usage{
beta.updater(X, Y, beta)
}
\arguments{
\item{X}{a design matrix. \code{X} must be of numeric type.}

\item{Y}{a response vector. \code{Y} must consist of \code{1} and \code{0} with \code{1} representing the event.
The length of \code{Y} must be equal to the number of rows in \code{X}.}

\item{beta}{a vector of the parameter values. \code{beta} must be of numeric type.
The length of \code{beta} must be equal to the number of columns in \code{X}.}
}
\value{
\code{beta.updater} returns a vector of the updated parameter values for IRLS method while fitting the logistic regression model.
The returned value has the same size as beta and is of numeric type.
}
\description{
\code{beta.updater} is used to update parameter values during IRLS iteration for fitting the logistic regression model.
\code{beta.updater} takes a design matrix, a binary response vector, and a vector of the current parameters.
}
\examples{
## Creating data from mtcars
dt <- mtcars
# Response: vs variable
Y <- dt$vs
# Covariates: mpg and am variables
X <- cbind(rep(1, 32), dt$mpg, dt$am)
# Initial beta
beta = rep(0,3)

## Calculate the updated parameters
beta.updater(X=X, Y=Y, beta=beta)

}
