% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1.loglik.R
\name{d1.loglik}
\alias{d1.loglik}
\title{Finding a Value of the First Derivative of the Log-Likelihood}
\usage{
d1.loglik(X, Y, beta)
}
\arguments{
\item{X}{a design matrix. \code{X} must be of numeric type.}

\item{Y}{a response vector. \code{Y} must consist of \code{1} and \code{0} with \code{1} representing the event.
The length of \code{Y} must be equal to the number of rows in \code{X}.}

\item{beta}{a vector of the parameter values. \code{beta} must be of numeric type.
The length of \code{beta} must be equal to the number of columns in \code{X}.}
}
\value{
\code{d1.loglik} returns a value of the first derivative of the log likelihood for the logistic regression model
The returned value is of numeric type
}
\description{
\code{d1.loglik} is used to find a value of the first derivative of the log-likelihood for the logistic regression model.
\code{d1.loglik} takes a design matrix, a binary response vector, and a vector of the parameters.
}
\examples{
## Creating data from mtcars
dt <- mtcars
# Response: vs variable
Y <- dt$vs
# Covariates: mpg and am variables
X <- cbind(rep(1, 32), dt$mpg, dt$am)
# Initial beta
beta = rep(0,3)

## Calculate the first derivative of log likelihood
d1.loglik(X=X, Y=Y, beta=beta)


}
